Experimental: tokenizers with and without templates
